Batch compare: initial version.
This is a new Occam mode, specialized for use in
Teresa Schmidt's PhD thesis.

The basic idea is, for many pairs of data files,
to run searches to find the "best model" for each file in the pair,
then to compare the two "best" models using a distance function.
In addition, output some other info (H, dDF, dAIC, dBIC).

This is not yet complete (but much of the "skeleton" is there).