Cached data mode (initial).
Marcus Harris often has to upload huge data files,
and got tired of waiting for the upload
after every minor change to the variable declaration block.

This mode allows combining separate decl-block and data-block files,
and produces a key that can be used to lookup a previous data-block,
to avoid re-uploading it.

Not yet finished or fully tested.