Added a shortcut to install Occam.
This is similar to Joe's approach
of installing to "../html",
but broken up into a few more steps.

I use a Makefile in the parent ("..") directory,
for an example see Makefile.parent,
and symlinks pointing into "~/public_html" to install Occam for testing.

Arguably this is the same approach the earlier Makefile used,
but I re-arrived at it just now, having familiarized myself
more with a development workflow...