Batch compare: enforce computation dependencies...
I'm not sure whether this commit is necessary.
At this point, the process of getting a fully-fitted Table
is arguably FAR TOO COMPLICATED:
many naive, but sensible-looking approaches that I tried
failed in various unintuitive ways.

We really need to make this kind of thing easier to do
with a single, well-named function ---
not an arcane series of invocations...