Removed unused `Readme` file.
It was sort of a "developer-facing" file ---
which I certainly didn't use.
Mostly, this seemed to be a listing of relevant functions,
without adding a whole lot of extra info.

I'm removing documentation here ---
but I found it confusing and unuseful,
and I think that confusing and unuseful documentation
is almost worse than no documentation,
since it could give developers (i.e. me!)
a false sense of security and cause them
to waste time on documents that don't give a full picture.

In this case, I think the code itself is the only "full picture"...
confusing as it is, it is also too stateful and implementation-dependent
to be easily described in terms of an API doc or diagram.
We need to improve this...
but while Martin obviously wants me to improve documentation,
he doesn't want me to prioritize it over features ---
so for now, I'm moving on...