Change style of 'include guards'.
This goes along with my other 'visual noise' commits...
I don't like the all-caps include guard names,
so I go for a style that's a little bit subtler ---
but unlikely to clash with any other defined name
due to the triple-underscore.

(GCC documentation says that double-underscore names are reserved for
the compiler, but says nothing about triple-underscore ---
presumably these are unlikely to clash with anything).