Neutral fit output: allow hiding complete table.
The complete Fit table is often many hundreds or thousands of lines
(and a whole lot of columns, too),
polluting the output with information that users often can't effectively
interpret.
By default, hide this table.

One goal is to add "Lift" computations that summarize the Neutral fit
in a more intuitive way.