Improve handling of cardinality errors in `Input.cpp`.
Specifically, a user ran into a problem with
a bizarre message about incorrect cardinality
which actually arose from a parse error
(due to an extra tab in a data line).

This error message doesn't preclude such an error recurring,
but it does make it clearer what is going on when it happens.

(Fixing the parser is another matter...)