Start removing static variables from Report.
It is not clear to me why these static variables were necessary,
but they're entrenched at this point.
Essentially, they are being used as global variables,
and really have nothing to do with any report instance.

I'd like to get rid of these global variables entirely if possible,
but for now I've tried to isolate them as much as I can.