Revert "Attempt to fix line counts."
This reverts commit f374c1d26369da0937d4ba8e86448e0f4c6fc46f.

This commit introduced an even worse error: incorrect sample size.
Actually, I think the original commit was fine,
but I mangled it when rebasing at a later time.

I think the error is due to inadvertently removing the
"values-are-functions" case-checking...
which was not intended to be part of this commit.

(Notably, Martin did actually ask me to disable the "function-values"
code for negative values, since it is unused and undocumented ---
but its use for probabilities is still important,
and not meant to be changed in the commit in concern).