BACK OUT zealous floating point accuracy checks for now.
These are causing some, hopefully mostly spurious,
issues with datasets that have "worked" in the past.

Temporarily removing these checks for now.

I suspect that most of the time, the floating
point issue has been simply underflow,
which is "harmless" in many places,
since a sufficiently low probability can be considered a 0 probability