Add (old) System Architecture document
Upload Ken Willett's document on the (proposed) OCCAM system architecture from 2000. The current version didn't totally follow the plan scheme laid out in this document (particularly with regard to implementing RA workflow in python - most of it is in c++). But this doc is valuable because, even though not all of its proposals were implemented, they are still valuable and can help inform planning for the structural changes that we are starting to plan.